services/frontend/requirements.txt
aio-pika==9.4.1
aiofiles==24.1.0
aiokafka==0.12.0
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.9.0
gradio==5.47.1
gradio_client==1.13.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pamqp==3.3.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
ruff==0.13.2
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.20.1
openpyxl